pyproject.toml
[tool.poetry]
name = "rubby"
version = "0.1.0"
description = ""
authors = ["Simon Fontaine <74004726+Simon-Fontaine@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
disnake = {extras = ["voice"], version = "^2.9.1"}
python-dotenv = "^1.0.0"
colorlog = "^6.8.0"
motor = "^3.3.2"
pydantic = {extras = ["email"], version = "^2.5.2"}
dateparser = "^1.2.0"
pendulum = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
